[feature](merge-cloud) compaction for mow table
