+++ be/src/olap/tablet.h
@@ -63,7 +63,6 @@ class CompactionMixin;
 class SingleReplicaCompaction;
 class RowsetWriter;
 struct RowsetWriterContext;
-class RowIdConversion;
 class TTabletInfo;
 class TabletMetaPB;
 class TupleDescriptor;
@@ -378,18 +377,7 @@ public:
                               DeleteBitmapPtr delete_bitmap, RowsetWriter* rowset_writer,
                               const RowsetIdUnorderedSet& cur_rowset_ids) override;
 
-    void calc_compaction_output_rowset_delete_bitmap(
-            const std::vector<RowsetSharedPtr>& input_rowsets,
-            const RowIdConversion& rowid_conversion, uint64_t start_version, uint64_t end_version,
-            std::set<RowLocation>* missed_rows,
-            std::map<RowsetSharedPtr, std::list<std::pair<RowLocation, RowLocation>>>* location_map,
-            const DeleteBitmap& input_delete_bitmap, DeleteBitmap* output_rowset_delete_bitmap);
     void merge_delete_bitmap(const DeleteBitmap& delete_bitmap);
-    Status check_rowid_conversion(
-            RowsetSharedPtr dst_rowset,
-            const std::map<RowsetSharedPtr, std::list<std::pair<RowLocation, RowLocation>>>&
-                    location_map);
-
     bool check_all_rowset_segment();
 
     void update_max_version_schema(const TabletSchemaSPtr& tablet_schema);
